requirements/base.txt
Django==5.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
djangorestframework==3.14
requests==2.31.0
gunicorn